LoM move
LoM
- New function available under Conquest Options to "Move to defend LoM
lands." If checked, it will automatically move you to the LoM land that
will defend next, unless you are already in a land that will be in
defend during the same time as that land.

Monster
- Increased Dashboard update calls for Monster Finder and Monster
- Fix unlimited healing when monster is in bonus to heal or strengthen
mode
- Fix to allow deflecting and crippling when When Fortify set to not
give energy for Fortify. To disable deflect and cripple, set
Cripple/Deflect When to "Never"
- Fix monster name only giving partial name if it has a space in it
- Fix Battle of the Dark Legion not finding fortification bar
- Fix changing monster selection string not updating monster max/ach
setting

Finder
- Reprioritize Monster Finder to be above passive general, so you can
use a monster general for joining new monsters.
- Fix energymax and staminamax to work in join :j conditions

Other
- Fix general refresh button not appearing when changing to general
dashboard